cifar10-image-classifier/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
--extra-index-url https://download.pytorch.org/whl/cpu